Move parseDatabaseValue to deserialization helper
Also fixed the case where an input would already have the Z.
