+++ apps/astro-app/src/utils/data/deserialization.ts
@@ -1,0 +1,100 @@
+import type { Datatype } from "@/types/data";
+import type { JSONValue } from "@utils/http";
+
+/**
+ * Attempts to parse a database value, expecting the given datatype. Throws an error if unsuccessful.
+ * @TODO integer verification.
+ * @param value The value that needs to be parsed.
+ * @param datatype @type {Datatype} The expected datatype of the value.
+ * @returns @type {string | number | boolean | Date} The date returned is UTC.
+ */
+function parseDatabaseValue(
+  value: JSONValue,
+  datatype: "int" | "integer" | "float" | "number",
+): number;
+function parseDatabaseValue(
+  value: string,
+  datatype: "string" | "str" | "text",
+): string;
+function parseDatabaseValue(
+  value: JSONValue,
+  datatype: "bool" | "boolean",
+): boolean;
+function parseDatabaseValue(
+  value: JSONValue,
+  datatype: "date" | "time" | "timestamp",
+): Date;
+function parseDatabaseValue(value: JSONValue, datatype: "enum"): string;
+function parseDatabaseValue(
+  value: JSONValue,
+  datatype: "geodetic point",
+): string;
+function parseDatabaseValue(
+  value: JSONValue,
+  datatype: Datatype,
+): string | number | boolean | Date;
+function parseDatabaseValue(
+  value: JSONValue,
+  datatype: Datatype,
+): string | number | boolean | Date {
+  switch (datatype) {
+    case "string":
+    case "str":
+    case "text":
+      if (typeof value != "string")
+        throw `Datatype mistmatch. Expected string but received ${typeof value}.`;
+      return value;
+    case "int":
+    case "integer":
+    case "float":
+    case "number":
+      switch (typeof value) {
+        case "number":
+        case "bigint":
+          return value;
+        case "string":
+          const result = parseInt(value);
+          if (isNaN(result))
+            throw `Failed to parse database value "${value}". Expected a number or number-like string.`;
+          return result;
+        default:
+          throw `Datatype mismatch. Expected number or number-like string but received ${typeof value}.`;
+      }
+    case "bool":
+    case "boolean":
+      switch (typeof value) {
+        case "boolean":
+          return value;
+        case "string":
+          switch (value.toLowerCase()) {
+            case "true":
+              return true;
+            case "false":
+              return false;
+            default:
+              throw `Failed to parse database value "${value}". Expected boolean or boolean-like string.`;
+          }
+        default:
+          throw `Datatype mismatch. Expected boolean or boolean-like string but received ${typeof value}.`;
+      }
+    // The database always stores UTC values.
+    case "date":
+    case "time":
+    case "timestamp":
+      switch (typeof value) {
+        case "string":
+          const result = new Date(value.endsWith("Z") ? value : value + "Z");
+          if (isNaN(result.getDate()))
+            throw `Failed to parse database value "${value}". Expected a ${datatype}-like string.`;
+          return result;
+        default:
+          throw `Datatype mismatch. Expected a ${datatype}-like string but received ${typeof value}.`;
+      }
+    case "enum":
+      return ""; // @TODO
+    case "geodetic point":
+      return ""; // @TODO
+  }
+}
+
+export { parseDatabaseValue };
